changed numeric table input to convert to float
to accommodate float values for fs, might need to change slightly if disp doesn't like floats

+++ napeca/gui.py
@@ -340,7 +340,7 @@ class MainWindow(QMainWindow):
                     file_tmp_dict[column_name] = False
 
                 if column_name in ['max_disp_y', 'max_disp_x', 'fs']:
-                    file_tmp_dict[column_name] = int(self.model_fparam_table.item(row, col).text())
+                    file_tmp_dict[column_name] = float(self.model_fparam_table.item(row, col).text())
 
             file_tmp_dict = combine_displacements(file_tmp_dict)
